Provide material detail for material_request endpoints.
