+++ api/resources_portal/serializers/__init__.py
@@ -1,0 +1,5 @@
+"""Serializers should only be added to this module instead of
+resources_portal.views if it is necessary to prevent a cyclical
+dependency."""
+
+from resources_portal.serializers.material import MaterialDetailSerializer, MaterialSerializer
